feat: add some lang utils and tests
