+++ jest.config.ts
@@ -7,13 +7,15 @@ const config: Config.InitialOptions = {
   setupFilesAfterEnv: ['<rootDir>/src/setupTests.ts'],
   transform: {
     '^.+\\.(js|jsx|ts|tsx)$': 'ts-jest',
-    '^.+\\.(css)$': 'jest.transform.js',
   },
   moduleNameMapper: {
     '.+\\.(css|styl|less|sass|scss|png|jpg|ttf|woff|woff2)$':
       'jest-transform-stub',
     'swiper/react': 'swiper/react/swiper-react.js',
     'swiper/css': 'swiper/swiper.min.css',
+  },
+  snapshotFormat: {
+    printBasicPrototype: true,
   },
 };
